user-service/requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
asyncpg
psycopg2-binary
kafka-python
python-dotenv

